[#35] refactor/dto 레이어를 이전처럼 dto 통합으로 관리
