+++ src/main/java/com/anipick/backend/anime/dto/AnimeIdTitleImgItemDto.java
@@ -1,12 +1,12 @@
-package com.anipick.backend.anime.common.dto;
+package com.anipick.backend.anime.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class AnimeItemDto {
-	private Long id;
+public class AnimeIdTitleImgItemDto {
+	private Long animeId;
 	private String title;
 	private String coverImageUrl;
 }
