+++ src/main/java/com/anipick/backend/anime/dto/AnimeItemDto.java
@@ -1,12 +1,12 @@
-package com.anipick.backend.anime.service.dto;
+package com.anipick.backend.anime.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class AnimeIdTitleImgItemDto {
-	private Long animeId;
+public class AnimeItemDto {
+	private Long id;
 	private String title;
 	private String coverImageUrl;
 }
