+++ src/main/java/com/anipick/backend/anime/dto/ComingSoonPageDto.java
@@ -1,6 +1,5 @@
-package com.anipick.backend.anime.controller.dto;
+package com.anipick.backend.anime.dto;
 
-import com.anipick.backend.anime.service.ComingSoonItemDto;
 import com.anipick.backend.common.dto.CursorDto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
