+++ src/main/java/com/anipick/backend/anime/dto/RangeDateRequestDto.java
@@ -1,4 +1,4 @@
-package com.anipick.backend.anime.service.dto;
+package com.anipick.backend.anime.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
