+++ src/test/java/com/anipick/backend/review/service/ReviewServiceTest.java
@@ -1,7 +1,7 @@
 package com.anipick.backend.review.service;
 
-import com.anipick.backend.review.service.dto.RecentReviewItemDto;
-import com.anipick.backend.review.controller.dto.RecentReviewPageDto;
+import com.anipick.backend.review.dto.RecentReviewItemDto;
+import com.anipick.backend.review.dto.RecentReviewPageDto;
 import com.anipick.backend.review.mapper.RecentReviewMapper;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
